refactor: filter duplicated messages
